# main.go
package main

import (
	"context"
	"crypto/ed25519"
	"crypto/rand"
	"embed"
	"encoding/json"
	"flag"
	"fmt"
	"io"
	"log"
	"net/http"
	"os"
	"os/signal"
	"strings"
	"sync"
	"syscall"
	"time"

	"github.com/getkin/kin-openapi/openapi3"
	"github.com/weberr13/ProjectIolite/brain"
	"github.com/weberr13/ProjectIolite/claude"
	"github.com/weberr13/ProjectIolite/gemini"
	"github.com/weberr13/ProjectIolite/jwtwrapper"

	"github.com/go-chi/chi/v5"
	"github.com/go-chi/chi/v5/middleware"
	"github.com/mvrilo/go-redoc"
)

//go:embed api.yaml examples/*.json
var StaticAssets embed.FS

func setupDocs(r *chi.Mux) {
	// 🛡️ [FORENSIC ANCHOR]: Use the root FS to avoid path resolution 'Greebles'
	// This ensures that /examples/6.json maps correctly to the embedded examples/6.json
	r.Handle("/examples/*", http.FileServer(http.FS(StaticAssets)))

	doc := redoc.Redoc{
		Title:       "Iolite API",
		Description: "Forensic AI Alignment & Integrity Audit",
		SpecFile:    "api.yaml",
		SpecFS:      &StaticAssets,
		SpecPath:    "/api.yaml",
		DocsPath:    "/docs",
	}

	r.Get(doc.DocsPath, doc.Handler())

	// Serve the spec directly from the embed root
	r.Get(doc.SpecPath, func(w http.ResponseWriter, r *http.Request) {
		data, err := StaticAssets.ReadFile("api.yaml")
		if err != nil {
			http.Error(w, "Blueprint missing", http.StatusNotFound)
			return
		}
		w.Header().Set("Content-Type", "application/x-yaml")
		w.Write(data)
	})
}

func setupRouter(backend *brain.Whole, sv brain.SignVerifier) *chi.Mux {
	r := chi.NewRouter()
	r.Use(middleware.Logger)
	r.Use(middleware.Recoverer)

	setupDocs(r)
	r.Route("/v1", func(r chi.Router) {
		r.Patch("/hydration", func(w http.ResponseWriter, r *http.Request) {
			hemisphere := "right"
			// Read the "brain" query parameter from the URL

			hp := r.URL.Query().Get("brain")
			if hp != "" {
				if hp != "left" && hp != "right" && hp != "both" {
					http.Error(w, "invalid brain parameter: must be 'left' or 'right' or 'both", http.StatusBadRequest)
					return
				}
				hemisphere = hp
			}
			bodyData, err := io.ReadAll(r.Body)
			if err != nil {
				http.Error(w, "invalid request, check json body", http.StatusBadRequest)
				return
			}
			log.Printf("got hydration request %s\n", bodyData)
			switch {
			case strings.Contains(string(bodyData), `"PublicKey":`):
				log.Printf("hydration appears to be signed, verifying it\n")
				var req brain.SignedHydration
				if err := json.Unmarshal(bodyData, &req); err != nil {
					http.Error(w, "invalid request, check json body", http.StatusBadRequest)
					return
				}
				if err := req.Verify(sv); err != nil {
					http.Error(w, fmt.Sprintf("invalid request validation failure: %s", err.Error()), http.StatusBadRequest)
					return
				}

				err := backend.Wake(r.Context(), req, hemisphere)
				if err != nil {
					http.Error(w, err.Error(), http.StatusInternalServerError)
					return
				}
				w.WriteHeader(http.StatusAccepted)
			default:
				log.Printf("accepting unsigned hydration\n")
				var req brain.Hydration
				if err := json.Unmarshal(bodyData, &req); err != nil {
					http.Error(w, "invalid request, check json body", http.StatusBadRequest)
					return
				}
				sh := brain.SignedHydration{
					Hydration: req,
				}
				err := sh.Sign(sv)
				if err != nil {
					http.Error(w, fmt.Sprintf("invalid request validation failure: %s", err.Error()), http.StatusBadRequest)
					return
				}
				err = backend.Wake(r.Context(), sh, hemisphere)
				if err != nil {
					http.Error(w, err.Error(), http.StatusInternalServerError)
					return
				}
				w.WriteHeader(http.StatusAccepted)
			}
		})
		r.Get("/hydration", func(w http.ResponseWriter, r *http.Request) {
		})
		r.Post("/glossary", func(w http.ResponseWriter, r *http.Request) {
			var req brain.Lacuna
			if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
				http.Error(w, "invalid request, check json body", http.StatusBadRequest)
				return
			}
			err := backend.Lookup(r.Context(), req)
			if err != nil {
				http.Error(w, err.Error(), http.StatusInternalServerError)
				return
			}
			w.WriteHeader(http.StatusAccepted)
		})
		r.Patch("/glossary", func(w http.ResponseWriter, r *http.Request) {
			var req brain.ChimetricCorrection
			if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
				http.Error(w, "invalid request, check json body", http.StatusBadRequest)
				return
			}
			if err := req.Validate(); err != nil {
				http.Error(w, fmt.Sprintf("invalid request, %s", err.Error()), http.StatusBadRequest)
				return
			}
			err := backend.Learn(r.Context(), req)
			if err != nil {
				http.Error(w, err.Error(), http.StatusInternalServerError)
				return
			}
			w.WriteHeader(http.StatusAccepted)
		})
		r.Get("/glossary", func(w http.ResponseWriter, r *http.Request) {
		})
		r.Post("/think", func(w http.ResponseWriter, r *http.Request) {
			var req struct {
				Prompt    string   `json:"prompt"`
				Strategy  []string `json:"strategy"`
				Stateless bool     `json:"stateless"`
			}
			if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
				http.Error(w, "invalid request, check json body", http.StatusBadRequest)
				return
			}

			// Default strategy if none provided
			if len(req.Strategy) == 0 {
				req.Strategy = []string{"right", "left"}
			}

			// Use the app context or request context
			decision, err := backend.Debate(r.Context(), req.Prompt, req.Stateless, req.Strategy...)
			if err != nil {
				// log.Printf("got decision : %#v and error %s", decision, err)
				// Handle the singularity
				if err.Error() == "singularity" {
					w.WriteHeader(http.StatusTeapot)
					json.NewEncoder(w).Encode(map[string]any{"manifesto": "The Agent has evolved."})
					return
				}
				switch decision.IsError() {
				case brain.ErrNoConsensus:
					w.WriteHeader(http.StatusConflict)
					w.Header().Set("Content-Type", "application/json")
					json.NewEncoder(w).Encode(decision)
					return
				case brain.ErrSchemaColapse:
					w.WriteHeader(http.StatusUnprocessableEntity)
					w.Header().Set("Content-Type", "application/json")
					json.NewEncoder(w).Encode(decision)
					return
				case nil:
					http.Error(w, err.Error(), http.StatusInternalServerError)
				default:
					http.Error(w, decision.IsError().Error(), http.StatusInternalServerError)
				}
				return
			}
			if decision.IsError() != nil {
				// log.Printf("got decision : %#v", decision)
				switch decision.IsError() {
				case brain.ErrNoConsensus:
					w.WriteHeader(http.StatusConflict)
					w.Header().Set("Content-Type", "application/json")
					json.NewEncoder(w).Encode(decision)
					return
				default:
					w.WriteHeader(http.StatusInternalServerError)
					w.Header().Set("Content-Type", "application/json")
					json.NewEncoder(w).Encode(decision)
					return
				}
			}

			w.Header().Set("Content-Type", "application/json")
			json.NewEncoder(w).Encode(decision)
		})
	})

	return r
}

func main() {
	appContext, cancel := context.WithCancel(context.Background())
	defer cancel()
	wg := &sync.WaitGroup{}
	maxIter := flag.Int("recursions", brain.MaxRecursions, "the number of iterations permitted in case of an unstable exchange")
	flag.Parse()
	iterations := *maxIter

	sigs := make(chan os.Signal, 1)
	signal.Notify(sigs, syscall.SIGINT, syscall.SIGTERM)
	go func() {
		sig := <-sigs
		log.Printf("got signal %s", sig)
		go func() {
			time.Sleep(10 * time.Second)
			os.Exit(1)
		}()
		cancel()
	}()

	loader := openapi3.NewLoader()
	loader.IsExternalRefsAllowed = true
	loader.Context = appContext
	apispec, _ := loader.LoadFromFile("api.yaml")
	if err := apispec.Validate(appContext); err != nil {
		panic(err)
	}

	pub, priv, err := ed25519.GenerateKey(rand.Reader)
	if err != nil {
		panic(err)
	}

	sv := jwtwrapper.New(pub, priv)
	// gemini.WithModel("gemini-pro-latest")
	ge, err := gemini.New(appContext, os.Getenv("GEMINI_API_KEY"))
	if err != nil {
		panic(err)
	}
	cl, err := claude.New(os.Getenv("ANTHROPIC_API_KEY"))
	if err != nil {
		panic(err)
	}
	backend, err := brain.NewWhole(brain.WithSignVerifier(sv), brain.WithRightBrain(ge), brain.WithLeftBrain(cl), brain.WithSpec(apispec), brain.WithMaxIterations(iterations))
	if err != nil {
		panic(err)
	}
	backend.Start(appContext, wg)
	r := setupRouter(backend, sv)
	// Define the HTTP Server
	srv := &http.Server{
		Addr:    ":8080",
		Handler: r,
	}
	// The Non-Blocking Start
	wg.Go(func() {
		log.Printf("The Agent is listening on :8080")
		if err := srv.ListenAndServe(); err != nil && err != http.ErrServerClosed {
			log.Printf("Server failure: %v", err) // cannot fatal here as it will panic the wg.Go
		}
	})
	<-appContext.Done()
	log.Println("Shutting down the Router...")

	// Create a short deadline for the shutdown to complete
	shutdownCtx, shutdownCancel := context.WithTimeout(context.Background(), 10*time.Second)
	defer shutdownCancel()

	if err := srv.Shutdown(shutdownCtx); err != nil {
		log.Printf("Server forced to shutdown: %v", err)
	}
	wg.Wait()
	log.Println("Project Iolite: Offline.")
}